rename variables, change fylesystem structure
