+++ src/auth/middleware/functions/checkIfTokenIsValid.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import passport from "../passport/passport";
-import { createResponse } from "../../responses/createResponse";
+import { createResponse } from "../../../responses/createResponse";
 
 export default function checkIfTokenIsValid(
   req: Request,
