+++ src/auth/middleware/functions/refreshAccessToken.ts
@@ -1,9 +1,9 @@
 import { NextFunction, Request, Response } from "express";
 import * as jwt from "jsonwebtoken";
 import passport from "../passport/passport";
-import { createResponse } from "../../responses/createResponse";
+import { createResponse } from "../../../responses/createResponse";
 
-export default function refreshAccessToken(
+export function refreshAccessToken(
   req: Request,
   res: Response,
   next: NextFunction
